feat: Server side swift (#680)
* Add linux target specific import of FoundationNetworking.
* Add Dockerfile for building the library in a container.
* Add SwiftCrypto as dependency for the Linux target.
* Operation related fixes for Linux.
* Import FoundationNetworking for the Linux target
* Add optional SwiftCrypto dependency for Linux targets to keep the compatibility with older macOS versions
* Implement integration tests skipping in case of missing environment variables

+++ Tests/AlgoliaSearchClientTests/Integration/OnlineTestCase.swift
@@ -33,19 +33,18 @@ class OnlineTestCase: XCTestCase {
     let dateString = dateFormatter.string(from: .init())
     return "swift_\(dateString)_\(NSUserName().description)"
   }
+  
+  var environment: TestCredentials.Environment { .default }
 
   override func setUpWithError() throws {
 
     try super.setUpWithError()
 
-    guard let credentials = TestCredentials.environment else {
-      throw Error.missingCredentials
-    }
-	
+    let fetchedCredentials = Result(catching: { try TestCredentials(environment: environment) }).mapError { XCTSkip("\($0)") }
+    let credentials = try fetchedCredentials.get()
+    
     client = SearchClient(appID: credentials.applicationID, apiKey: credentials.apiKey)
-    let className = String(reflecting: type(of: self)).components(separatedBy: ".").last!
-    let functionName = invocation!.selector.description
-    let indexNameSuffix = self.indexNameSuffix ?? "\(className).\(functionName)"
+    let indexNameSuffix = self.indexNameSuffix ?? name
     let indexName = IndexName(stringLiteral: "\(uniquePrefix())_\(indexNameSuffix)")
     index = client.index(withName: indexName)
 
@@ -70,7 +69,9 @@ class OnlineTestCase: XCTestCase {
 
   override func tearDownWithError() throws {
     try super.tearDownWithError()
-    try index.delete()
+    if let index = index {
+      try index.delete()
+    }
   }
 
 }
